automake: don't delete .pyc files
The patch being removed in this commit removes *.pyc files from being
compiled. This dose not allow a user to select which files are included
in the image. Since optimization is no longer the default for python
we should have the ability to choose what is included in the image.

(From OE-core rev: bee0f633f0b5ef2439b8e5d54cbfc7bc1f17f47e)

Signed-off-by: Joseph A. Lutz <joseph.lutz@novatechweb.com>
Signed-off-by: Ross Burton <ross.burton@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>